Moving the file translator to a file lookup
