+++ stacker/lookups/registry.py
@@ -4,6 +4,7 @@ from ..util import load_object_from_string
 from .handlers import output
 from .handlers import kms
 from .handlers import xref
+from .handlers import file as file_handler
 
 LOOKUP_HANDLERS = {}
 DEFAULT_LOOKUP = output.TYPE_NAME
@@ -53,3 +54,4 @@ def resolve_lookups(lookups, context, provider):
 register_lookup_handler(output.TYPE_NAME, output.handler)
 register_lookup_handler(kms.TYPE_NAME, kms.handler)
 register_lookup_handler(xref.TYPE_NAME, xref.handler)
+register_lookup_handler(file_handler.TYPE_NAME, file_handler.handler)
